+++ src/components/manager/Responsive/ResponsiveProvider.js
@@ -1,5 +1,5 @@
-import React, { Component, Children } from 'react';
-import { element, shape, number } from 'prop-types';
+import React, { Component } from 'react';
+import { element, shape, number, func, bool } from 'prop-types';
 
 import throttle from 'lodash.throttle';
 import { getQueries } from './getQueries';
@@ -8,11 +8,14 @@ import { defaultBreakpoints } from './breakpoints';
 
 const fakeWindow = { innerWidth: 0, innerHeight: 0 };
 
+const getContainer = ({ getContainer, container }) =>
+  typeof getContainer === 'function' ? getContainer() : container;
+
 export class ResponsiveProvider extends Component {
   constructor(props) {
     super(props);
     this.state = this.calculate(props);
-    this.listen = throttle(this.updateSizes, props.throttle);
+    this.listener = throttle(this.updateSizes, props.throttle);
   }
 
   updateSizes = () => {
@@ -20,22 +23,24 @@ export class ResponsiveProvider extends Component {
   };
 
   calculate(props) {
-    return getQueries(props.container, props.breakpoints);
+    return getQueries(getContainer(props), props.breakpoints);
   }
 
   componentDidMount() {
-    this.props.container.addEventListener('resize', this.listen);
-    this.updateSizes();
+    if (this.props.runOnMount) {
+      this.updateSizes();
+    }
+    getContainer(this.props).addEventListener('resize', this.listener);
   }
 
   componentWillUnmount() {
-    this.props.container.removeEventListener('resize', this.listen);
+    getContainer(this.props).removeEventListener('resize', this.listener);
   }
 
   render() {
     return (
       <ResponsiveContext.Provider value={this.state}>
-        {Children.only(this.props.children)}
+        {this.props.children}
       </ResponsiveContext.Provider>
     );
   }
@@ -44,6 +49,7 @@ export class ResponsiveProvider extends Component {
 ResponsiveProvider.propTypes = {
   children: element.isRequired,
   container: shape({}),
+  getContainer: func,
   breakpoints: shape({
     xs: number.isRequired,
     sm: number.isRequired,
@@ -52,6 +58,7 @@ ResponsiveProvider.propTypes = {
     xl: number.isRequired,
     xll: number.isRequired,
   }),
+  runOnMount: bool,
   throttle: number,
 };
 
@@ -59,4 +66,5 @@ ResponsiveProvider.defaultProps = {
   container: typeof window === 'undefined' ? fakeWindow : window,
   breakpoints: defaultBreakpoints,
   throttle: 400,
+  runOnMount: false,
 };